chore | AWS 배포 관련 설정 및 GitHub Actions 관련 파일 추가
