+++ .github/workflows/deploy.yml
@@ -1,0 +1,36 @@
+name: Where Is My Hometown 
+on: 
+  push:
+    branches:
+      - main 
+jobs: 
+  build:
+    runs-on: ubuntu-20.04 
+    steps:
+      - name: Git Checkout 
+        uses: actions/checkout@v2
+
+      - name: Check Node Version 
+        run: node -v
+
+      - name: Install Dependencies
+        run: |
+          npm install -g yarn
+          yarn install --frozen-lockfile
+
+      - name: Build
+        run: yarn build
+
+      - name: Configure AWS credentials 
+        uses: aws-actions/configure-aws-credentials@v1
+        with:
+          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }} 
+          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }} 
+          aws-region: ${{ secrets.AWS_REGION }} 
+
+      - name: Deploy to S3 
+        run: | 
+          aws s3 sync ./${{ secrets.BUILD_DIRECTORY }} ${{ secrets.AWS_S3_BUCKET_NAME }} --acl public-read --delete
+
+      - name: CloudFront Invalidate Cache
+        run: aws cloudfront create-invalidation --distribution-id ${{ secrets.AWS_CLOUDFRONT_DISTRIBUTION_ID }} --paths '/*'
